Add cacheable param for fields and attr for model
For a field, just pass `cacheable=False` in the field declaration to
disable caching for this field.

For a model, just add a `cacheable = False` attribute to the class
defining the model fr which you don't want cache.

A model with `cacheable` to False implies all fields with `cacheable` to
False