Subclasses models do not share collections names
There was a bug with the _parent_class attribute of fields which was set
on fields, because fields were shared between model classes.

Ex:

```
class A(RedisModel):
    a = StringField(indexable=True)

class B(A):
    b = StringField(indexable=True)
```

Previsouly the `a` field was indexed using the same collection for both
models.
This is tested in `InheritanceTest.test_inheritance_collections`.