Implement a RedisDatabase object
This model is mandatory on each model, and can be defined in a base
model.
This new database manage connection, and the pipeline command (which
does not much more than the default redis pipeline:
- saving the normal connection
- using the new pipeline connection everywhere
- saving a state "pipeline_mode"

This pipeline is not yet used for adding atomicity inside limpyd, and
does not work well with cache and indexes. DO NOT USE IT ;)