Add a way do declare models as abstract
It's done by adding an "abstract" attribute, default to False if not
defined.
It's actually set to True on RedisModel and can be set on every of its
subclasses to avoid checking for database and namespace.
With this, it's easier to define some of these subclasses in contrib for
other uses.