Split (de)index in (de)index_value and (de)index
The new `index_value` and `deindex_value` methods handle all the logic
that was previously done in `index` and `deindex`, but need a value to
work with.
This value is given by the `index` and `deindex` methods which get it
from `proxy_get`.

The aim is to allow other fields to be indexable, without restricting
indexable fields to have only one value, or to (de)index a modified
value.