Add methods to attach a field to a model/instance
The previous way to say that a field was attached to a model or later to
an instance was to do:
    field._model = model
or
    field._instance = instance

But in RedisField subclasses it would be great to be able to catch this to
do some stuff. With this commit, it's as easy as to override the new
_attach_to_model and _attach_to_instance methods, which actually simply
do the assignments as before.