Add relations between related fields and models.
All RelatedField now attend a "to" parameter to define which model is
related, and a "related_name" (optional) to define the name of the
reverse relation.

Example:

    class Person(RelatedField):
        name = HashableField()
        group = FKHashableField('Group', related_name='members')

    class Group(RelatedField):
        name = HashableField()

    group = Group(name='my group')
    person = Person(name='me', group=group)

    members = Person.collection(group=group._pk)
    # or
    members = group.members()

The "to" parameter can be:
  - a model
  - the string "self" to design the current model
  - the name of a model (implies the namespace of the current model)
  - the full name ('namespace:name') of a model

The "related_name" parameter can be:
  - not defined, and it will be: "%s_set" % current_field.name
  - a string, which must be a valid python identifier
  - a formatable string, with %(namespace)s and/or %(model)s which will
    be replaced by the namespace and model of the current model (usefull
    for an abstract model)

The related_name si used to create a shortcut to a collection, as shown
with "members =" lines above. You can pass additional filters, as for a
real collection (in fact it is/returns a real collection)