Do not check existence of new AutoPKField value
To get a new pk from a AutoPKField, we increment the key where the last
one is used. As it's atomic, we are sure this new value was never used
before (except if someone played manually...)
So, in the case of AutoPKField, we don't need to check if this pk is
already used.
Result: one redis call less for each new object with AutoPKField