No existence check when setting default values.
As the `set_defaults` method is called in `get_pk`, just after we have a
brand new pk for the instance, we can assume that no fields are saved
yet, and so save the fields with default values without checking if
these fields already had a value. They should not.
I prefixed the `set_defaults` method with an underscore to indicate that
it's purely an internal method.
With this optimization, we have one redis call less for each default
value saved.