Change internal of RedisModel.exists
The exists method create a collection with its parameters, to check if
the result is empty or not.
The previous way was to retrieve the whole collection, couning the
number of elements, and return True if positive, or False if 0.
But this way, if the resulting collection returns a lot of instances, it
can cost a lot in bandwidth and cpu on the python side.
The new way resolves this by asking for only the first element (which
implies a sort, but we use "by nosort" in this case)
A little more work on the redis side, but a lot less bandwidth and cpu
on the python side.