Split available modifiers in full and partials
And now commands are all in a _commands attribute of the class:

_commands['all']: available commands
_commands['getters']: commands used to retrieve data from Redis
_commands['modifiers']: commands used to modify data in Redis
_commands['full_modifiers']: commands used to modify data in Redis with defined final value
_commands['partial_modifiers']: commands used to modify data in Redis, without knowing the final value.

The only entries that must be defined are 'getters', 'full_modifiers' and 'partial_modifiers'.
The others are computed based on these three ones.

We slit full modifiers and partial modifiers to handle differently the "transaction" for indexed fields.