Do big operations inside a lock
When a field is updated and needs deindexing/indexing, we run all the
needed operations inside a redis lock.
All the field is locked for writting, for all instances. Only the current
thread can do many operations on it.
The aim of this is to avoid collisions while updating indexes of a
field.