Make lock stuff threadsafe
To allow many operations during a single "command", which may include
deindexing and indexing, we store a flag if the field is currently locked or
not, to bypass lock for other operations within the main lock.

Before this commit, the flags where saved on each RedisModel, at the class
level.
But it was not threadsafe at all, a new thread could have bypassed the lock
because of this.

Now the flags are saved in threading.local