Add consistency when using _skip_exist_test
Before this commit, when using _skip_exist_test to get an object
(via the "instances" method of a collection or directly via the model
constructor), no check of the pk existence were ever made.

But when setting a field, entries are created and index updated. And
if a non-existent pk, the index contains a bad pk and retrieving can
fail. Lack of consitency.

With this commit, the existence of the pk will be done the first time
an update is done on an object if we first asked to skip it.

No check is done when only getters are called (to keep the speed
gained by using the skip parameter).

To do this, the _traverse_command is now surrounded by a _call_command
which become the main entry point to call commands.
_traverse_command is the one to override to update its behaviour, as
already done in our subclasses.