Make hmget and hmset use _traverse_command
To do this, i had to rewrote make_cache_key so it can handle dict
and list in args or as values of kwargs, recursively. So it can now
work with a lot of configurations.

I also had to cache management for RedisModel in the RedisModel itself
but also in memoize. (All cache, for the instance and its fields, is,
as previously, stored in the _cache attribute of the instance. To
avoid collision, cache for the instance is stored in the "__self__"
key while cache for fields use the fields names