hgmet & hmset can know accept args as in redis-py
In redis-py, hmget waits for a list of keys to retrieve, not *args (
which is not consistent with the rest of their api)
In limpyd, we waited for *args, not a list.
It's now possible to use the both ways in limpyd: a list as in
redis-py, or as *args.

It's the same for hmset, waiting for a dict in redis-py, and for
**kwargs in limpyd. Now both ways are available in limpyd.