Disable caching hmget result
We don't want to cache the result of hmget because if a hset is done
on a field, the hmget cache won't be cleared and a call will return
an old value.

So we now have a new type of "_commands", "no_cache_getters", with
commands that acts as a getter, but always ignoring the cacheable
flag.

I slightly rewrite MetaRedisProxy to correctly handle all the commands
types, to be sure we have sets, and to allow user to not declare an
empyt if it will be empty (+doc on this _commands attribute)