We can now use another CollectionManager
It can be useful to use another CollectionManager than the default one, to
active a filter by default (for example only return 'active' objects...),
or to use an enhanced manager (as we'll see later in contrib.collection)

There is two way to do this.

At the model level, by assigning a subclass of CollectionManager to the
"collection_manager" attribute:

    class MyModel(RedisModel):
        # [here fields and other stuff]
        collection_manager = MyCustomCollectionManager

The other way is to pass the wanted CollectionManager subclass to the
"manager" named argument to the "collection" method.

With this, in our previous example, we can still use the default manager:

    a_collection = MyModel.collection(manager=CollectionManager,
                                      a_filter=a_value[...])

So if yu want to keep the default manager on the model, you can use a
different one when calling the collection.