Add an ExtendedCollectionManager in contrib
Located in contrib.collection, this ExtendedCollectionManager extends the
defaut CollectionManager by provinding two new methods:
- filter
- intersect

"filter" simply calls the recently added "_add_filters", and add given
filters to the list of all filters to take into account. This method
returns the current collection, so its chainable, and can be used many
times.
All the filters are "and"ed, so if a filter is set with a value in a call
and with another value in another call, you can be sure that the collection
will be empty.

"intersect" takes a list of sets that will be intersected with the ones
got from the calls to "_add_filters". It can be the name of a redis set,
a SetField, or a regular set/list/tuple. This method returns the current
collection, so its chainable, and can be used many times.

All sets for all the calls to filter and intersect will be "intersected"
in redis during the call to the collection