Use scriping if available to copy list/zset to set
In ExtendedCollectionManager, the "intersect" method allow the use of a
ListField or SortedSetField.
But as they are not stored as a redis set, they cannot be used directly
to compute the collection (need to "intersect" the different sets)

Before this commit, the whole content of the field was retrieved locally
before its content put back in a temporary redis set. It's can take a long
time, and a lot of bandwith.

With this commit, this copy is done atomically in one redis call, using
the scripting capabilities offered by redis 2.6.
This, only if both the client (redis-py) and the server (redis) support
this feature, else we fall back to the previous way.