Use ZINTERSTORE when intersecting with sorted sets
The way to get results from a collection is to do call SINTERSTORE to
intersect on all set fields.
With the "intersect" method, we can have not only redis sets, but also
lists, tuples, sets, redis list and redis sorted set. All of these were
converted to set before calling SINTERSTORE.

But it's useful to know that redis can apply ZINTERSTORE not only on
sorted sets, but also on sets. So we don't need anymore to convert the
sorted sets passed to "intersect" (veru useful of the sorted set is a big
one).

We just have to call ZINTERSTORE instead of SINTERSTORE if we have at
least a sorted set. And on the execution, if no call to SORT is needed,
we call ZRANGE instead of SMEMBERS.