Allow sorting collections by a zset's score
Sadly, sorting by a sorted set's score is not possible in the "sort"
command in Redis, so to allow this, we have to convert the sorted set
in keys and sort using the '*' placeholder.

If we have a slice or want to get values, we have to do the convert before
the call to sort. It can be a big operation if the sorted set is huge.
If no slice nor values, we can do it after the call to redis (smembers or
sort) and only make keys for returned values.

Another possible optimisations is, when we do the convert at the begining,
is to only make keys for values matching the intersection of all sets. But
it's not for now ;)