Filters can accept fields on right part.
In CollectionManager, left part of a filter is the field's name, and
the right part is the value, used to get the key holding its index.
In ExtendedCollectionManager, the right part can also be a
SingleValueField (StringField, HashableField or PKField), and the value
will be fetched from Redis only when the collection is really called.