Get score in values/list if sorted by score
When sorting by score, and using values/values_list, it can be useful to
get the score from the sorted set used to sort for each value.
It's now possible with the collection manager in contrib, by passing the
constant "SORTED_SCORE" as argument to values or values_list, with the
other wanted fields.
If this constant (in contrib.collection) is used but there is not sort,
or no sort by score, the final value for this field will be None.