Add "store" method to ExtendedCollectionManager
It's now possible to store the result of a collection, using "store",
which take two optional arguments:

- `key`, which is the Redis_ key where the result will be stored, default
   to a randomly generated one
- `ttl`, the duration, in seconds, for which we want to keep the stored
   result in Redis_, default to `DEFAULT_STORE_TTL` (60 seconds, defined
  in `contrib.collection`). You can pass None if you don't want the key
  to expire in Redis_.

When calling `store`, the collection is executed and you got a new
ExtendedCollectionManager_ object, pre-filled with the result of the
original collection.

Note that only primary keys are stored, even if you called `instances`,
`values` or `values_list`. But arguments for these methods are set in the
new collection so if you call it, you'll get what you want (instances,
dictionaries or tuples). You can call `primary_keys` to reset this.

More informations in README.txt