Speed up "len" calls of collections.
When we directly ask for the len of a collection, we can bypass the call
to the sort method, and directly as redis the length of the "final set".

As we managed the "__iter__" + "__len__" problem (see d9283d4bfb: the
"__len__" is called when just after iterating), we have to handle two
cases: we ask for the len, or for iteration.

It's one by setting a new internal attribute "_len_mode" to True by
default, setting it to False when we want to iterate, and to True when
done. So when "__len__" is manually called, it's set to True, but when
it's called after a "__iter__", it's set to False