Add a way to change database on the fly
The main scenario is: an external lib define some models, but as the
database is mandatory, it may not be on the correct redis server/db.
And another problem, you cannot use these models on related fields
defined by another/your database because related fields must be on the
same database.

So this commit introduce the "use_database" method on RedisModel.
It's a class method, so all subclasses of the concerned model will
have the new database (only if they shared the same one as the parent
model)

Example. Say we have a lib with some models:

    class BaseModel(RedisModel):
        database = RedisDatabase()
        abstract = True

    class Foo(BaseModel):
        # ... fields ...

    class Bar(BaseModel):
        # ... fields ...

So in your code you can just do

    database = RedisDatabase(**connection_settings)
    BaseModel.use_database(database)

Note that you can use it the other way, via the "use_for_model" method
on the database (which is in fact called by "use_database"):

    database.use_for_model(BaseModel)