Allow usage of debugger in collection
This is done by restoring "state" (_slice and _len_mode) when quitting a
function.
When debugging, the debugger my call __len__ to display the actual
variables and then change one of these values, so the state of the
current debbuging is broken.
Note that it's still not thread-safe but still better than before.