Extract all index related stuff into a standalone class
This class, `EqualIndex` is used by default when the field is indexable.
It is based on `BaseIndex` that will be the base for other indexes to
come. Each index handle the indexing, deindexing, and the filtereing (by
returning a redis key that will then be used by the collection)

The new indexes handle a filter "suffix", for example, for the
EqualIndex: "None" (no suffix) or "__eq".

When such a suffix is encountered in a filter, the first index saying it
can handle it will do.

It's the same for uniqueness, each index says if it can or not handle
uniqueness, and then the first one capable of will check for uniqueness
(to avoid doing too much work if many indexes say thay can handle it: we
trust the first one)

For now, and maybe for later, the primary keys are still not indexable
(they are managed in a special way), but we added the ability to use the
prefix "__eq" for them too.

For the basic collection, we changed the moment of the index is called
to get the key: instead of the moment the filter is added, it's now when
the collection is called. It doesn't change a lot of things for
EqualIndex, but it may for more complex indexes that may have to do some
computation, locally or at the redis level.

Note that this commit does not change any existing test: it is 100%
compatible with existing usage of limpyd