Use lua in redis for range filtering
Previously, we fetched the result from the edis zrangebylex call,
extracted the primary keys, then sent them back to redis in a set or
sorted set.
The problems are:
- it could be a lot of data to transfer
- so a big cost in memory for the caller
- and it could also be a long time blocking redis to return the data

Now with the script, we simple tell redis to do this in lua: no data is
transfered to the caller.
And in the script, the fetch of data via zrangebylex is done by blocks,
so it won't take a lot of memory