Adds ability to override behavior without subclassing
Simple with a class method named `override` that can accept a prefix or
transform.

The prefix will allow to differentiate two same indexes, at the storage
level and the query level ("myfield__prefix=...")

The transform is a function that will transform the value to be stored.
For example to store only the year of a date.