Correct and enhance collection slicing
Some slicing were not managed at all, retrieving the wrong number of
data.
It is now corrected, with also more cases where we try to limit the data
returned by redis
To be sure that the data is always correctly returned, in the tests we
compare the result of slicing a collection to the same slicing of a
normal python list. Same for indexing (both in `__getitem__`)