Collections enhancements: generators + cloning
When calling `sort`, `instances`, `primary_keys`, `filter`, `values`,
and `values_list`, a new collection is returned instead of updating the
current one.

And when iterating over a collection, we iterate over a generator, not a
list (so we don't have to create all objects before getting the first
one: they will be created one by one)

And finally, when calling `instances`, the instances that would have
raised a `DoesNotExist` exception because of a PK in an index that does
not exist anymore, are now skipped.