Abstract code from `index`/`deindex` method from fields
Now the main work is done in `_index` and `_deindex` and does not need
to be repeated in subclasses.
The subclasses simply define `_prepare_index_data` and simple `index`
and `deindex` that will call `_index` and `_deindex`.

This also allows, via `_prepare_index_data`, to get the data to
index/deindex outside of the index/deindex methods.