Handle UniquenessError in race condition at create time
Fields are checked ony by one, then added one by one.
But if another instance is created between the check and the add, we'll
have a UniquenessError raised by the add, that won't be managed, so
we'll have an instance partially created.
Now, the instance will be deleted if this occurs.