Double layers cache for collection
- when calling `__len__` first, the redis set with keys is not deleted
and kept in case a collection retrieval is done just after, to avoid
having to recompute a possible complex final redis set. This key has an
expiration (300 seconds) and is automatically deleted when the
collection is fetched
- when the collection is fetched, data is locally cached, and calls like
`__len__`, `__getitem__`, `__iter__`... are applied on this cached data,
without the need to call redis again, which was the case if they were
applied on the collection itself and not a `list` of this collection

The goal was to avoid having to think in which order to make calls and
to not forget using `list` to not have too many redis calls.

Collections now also have a `__bool__`, `__repr__` and `__eq__` methods.

As before, calling `__getitem__` with a slice returns an iterator. But
if called with an index, it will raise an `IndexError`, as expected,and
not a `StopIteration` anymore.